requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
absl-py==2.2.2
addict==2.4.0
audioread==3.0.1
basicsr==1.4.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crayons==0.4.0
decorator==5.2.1
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
grpcio==1.71.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
jieba==0.42.1
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.0
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.43.0
lmdb==1.6.2
lpips==0.1.4
Markdown==3.8
MarkupSafe==3.0.2
mir_eval==0.8.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pesq==0.0.4
piexif==1.1.3
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.30.2
pycparser==2.22
pydub==0.25.1
pypdf==5.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
stegano==0.11.3
sympy==1.14.0
tb-nightly==2.20.0a20250511
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.67.1
#triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
pyyaml>=6.0
colorlog>=6.7.0
clip @ https://github.com/openai/CLIP/archive/refs/heads/main.zip